feat(ThemeContext): Implement  component
