+++ components/Speakers.js
@@ -1,7 +1,8 @@
 import React, { useContext } from "react";
 import SpeakersList from "./SpeakersList";
 import SpeakersToolbar from "./SpeakersToolbar";
-import { ThemeContext } from "./Layout";
+// import { ThemeContext } from "./Layout";
+import { ThemeContext } from "./contexts/ThemeContext";
 
 const Speakers = () => {
   const { showSessions } = useContext(ThemeContext);
